requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
langchainhub
beautifulsoup4
requests
docarray
python-dotenv